Solved LeetCode 66: Plus One
Challenging at first. Underestimated the challenge and would get the final test wrong because it would return 10 instead of 1,0

You are given a large integer represented as an integer array digits, where each digits[i] is the ith digit of the integer. The digits are ordered from most significant to least significant in left-to-right order. The large integer does not contain any leading 0's.

Increment the large integer by one and return the resulting array of digits.